+++ quizApp/src/app/display-quiz/display-quiz.component.html
@@ -4,6 +4,7 @@
   <div class="flex">
 
     <mat-card class="width70">
+      
       <mat-card-title class="center quizTitle">{{quiz.title}}</mat-card-title>
       <br>
       <div class="questionContainer">
